Edit Problem 1. Long to Int.

+++ intro-and-basic-syntax/exercise/Problem 1.cs
@@ -9,8 +9,8 @@ namespace _01._Debit
             // Initializing variables.
             int first = int.Parse(Console.ReadLine());
             int second = int.Parse(Console.ReadLine());
-            long third = long.Parse(Console.ReadLine());        //Edited
-            long forth = long.Parse(Console.ReadLine());        //Edited
+            int third = int.Parse(Console.ReadLine());        //Edited
+            int forth = int.Parse(Console.ReadLine());        //Edited
 
             //Printing result.
             Console.WriteLine($"{first:d4} {second:d4} {third:d4} {forth:d4}");
